Move inventory_toggle_button into src/, include/ folders
